+++ api/lookup.php
@@ -1,9 +1,22 @@
-<?php
-$data = [
-    "data1" => "1",
-    "data2" => "2",
-    "data3" => "3",
-    "data4" => "4",
-];
+<?php // Lookup patient details
 
-echo json_encode($data);
+/*
+Post patientId, hostpitalId and nhsNumber
+Return status, patientId, patientAge, firstName, lastName, dob, hospitalId, nhsNumber, patientSex
+Codes:
+    - 30: Request received but not yet processed
+    - 31: Request being processed
+    - 32: Request successful and patient data in response
+*/
+
+// Firstly, if we have no data, quit.
+if(!$data) {
+    // Throw error and stop processing things.
+    echo(json_encode(["Error" => "No data available."]));
+    exit;
+}
+
+$response = "API Unfinished.";
+
+// Provide the response
+echo json_encode(["status" => $response]);