+++ api/qccheck.php
@@ -1,9 +1,21 @@
-<?php
-$data = [
-    "data1" => "1",
-    "data2" => "2",
-    "data3" => "3",
-    "data4" => "4",
-];
+<?php // Check if the specified product and lot number has been QC tested.
 
-echo json_encode($data);
+/*
+Get productCode, lot, moduleSerialNumber
+Return status, isValid(bool)
+Codes:
+    - 50: Success
+*/
+
+// Firstly, if we have no data, quit.
+if(!$data) {
+    // Throw error and stop processing things.
+    echo(json_encode(["Error" => "No data available."]));
+    exit;
+}
+
+$response = "API Unfinished.";
+
+// Provide the response
+echo json_encode(["status" => $response]);
+