+++ views/assay-modules.php
@@ -1,0 +1,1 @@
+<h1>Assay Modules</h1>