+++ views/login.php
@@ -1,1 +1,6 @@
-<h1>Login</h1>
+<div class="flex flex-col justify-center items-center w-full grow">
+    <h1>Caleb is stinky</h1>
+    <p>And this is the login page...</p>
+</div>
+
+<p class="fixed top-0 left-0 pointer-events-none text-primary">Caleb is stinky</p>