+++ views/qc-policy.php
@@ -1,0 +1,1 @@
+<h1>Quality Control Policy</h1>