+++ views/versions.php
@@ -1,0 +1,1 @@
+<h1>buncha nonsense</h1>